# apiService/src/swagger.json
{
    "swagger":"2.0",
    "info":{
       "description":"This is the swagger node.js API project",
       "version":"1.0.0",
       "title":"Authentication API ES "
    },
...
    "host":"localhost:4000",
    "paths":{
...
"/driver/signUp":{
   "post":{
      "summary":"user signUp",
      "description":"creation of user",
      "produces":[
         "application/json"
      ],
      "parameters":[
         {
            "in":"body",
            "name":"user",
            "description":"user object",
            "required":true,
            "schema":{
                 "type":"object",
                 "$ref":"#/definitions/user"
                 
               
            }
         }
      ],
      "responses":{
         "201":{
            "description":"user created sccussfully",
            "sechema":{
               "type":"object",
                "$ref":"#/definitions/userResponse"
               
            }
         },
         "409":{
            "description":"user alredy exist",
            "sechema":{
               "$ref":"#/definitions/userResponse"
            }
         },
         "500":{
             "description":"something went wrong",
             "sechema":{
                "$ref":"#/definitions/userResponse"
             }
         }
      }
   }
},
"/driver/login":{
   "post":{
      "summary":"user login",
      "description":"login user ",
      "produces":[
         "application/json"
      ],
      "consumes":[
         "application/json"
      ],
      "parameters":[
         {
            "in":"body",
            "name":"body",
            "description":"userLogin object",
            "required":true,
            "schema":{
             "type":"object",
             "$ref":"#/definitions/userLogin"
               
         
         }
     }
      ],
      "responses":{
         "201":{
            "description":"user login is done",
            "schema":{
               "type":"object",
                  "$ref":"#/definitions/userResponse"
               
            }
         },
         "401":{
            "description":"Invalid Email or Password",
            "schema":{
               "$ref":"#/definitions/userResponse"
            }
         },
         "404":{
             "description":"user doesnot exist",
             "schema":{
                "$ref":"#/definitions/userResponse"
             }
          },
          "500":{
             "description":"something went wrong please try again",
             "schema":{
                "$ref":"#/definitions/InvalidResponse"
             }
          }
      }
   }
},
"/driver/createNewlist":{
   "post":{
      "summary":"creation of NewList",
      "description":"creation of NewList",
      "produces":[
         "application/json"
      ],
      "parameters":[
         {
            "in":"body",
            "name":"NewList",
            "description":"NewList object",
            "required":true,
            "schema":{
                 "type":"object",
                 "$ref":"#/definitions/newCar"  
            }
         }
      ],
      "responses":{
         "201":{
            "description":"NewList created sccussfully",
            "sechema":{
               "type":"object",
                "$ref":"#/definitions/userResponse"
               
            }
         },
         "500":{
             "description":"something went wrong",
             "sechema":{
                "$ref":"#/definitions/userResponse"
             }
         }
      }
   }
},
"/driver/addContactInfo":{
   "post":{
      "summary":"creation of addContactInfo",
      "description":"creation of addContactInfo",
      "produces":[
         "application/json"
      ],
      "parameters":[
         {
            "in":"body",
            "name":"addContactInfo",
            "description":"addContactInfo object",
            "required":true,
            "schema":{
                 "type":"object",
                 "$ref":"#/definitions/addContactInfo"  
            }
         }
      ],
      "responses":{
         "200":{
            "description":"ContactInfo created sccussfully",
            "sechema":{
               "type":"object",
                "$ref":"#/definitions/userResponse"
               
            }
         },
         "400":{
             "description":"something went wrong",
             "sechema":{
                "$ref":"#/definitions/userResponse"
             }
         }
      }
   }
}
    },
    "definitions":{
       "userResponse":{
          "type":"object",
          "properties":{
             "status":{
                "type":"integer"
             },
             "message":{
                "type":"string"
             }
             
          }
       },
...
       "user":{
          "type":"object",
          "properties":{
             "id":{
                "type":"objectId"
             },
             "userName":{
                "type":"string"
             },
             "email":{
                "type":"string"
             },
             "password":{
                "type":"string"
             },
             "phoneNumber":{
                "type":"string"
             },
             "address":{
               "type":"string"
             },
             "locality":{
               "type":"string"
             },
             "area":{
               "type":"string"
             },
             "zipcode":{
               "type":"string"
             },
             "city":{
               "type":"string"
             }

          }
        },
          "userLogin":{
            "type":"object",
            "properties":{
              
               "email":{
                  "type":"string"
               },
               "password":{
                  "type":"string"
               }
            }
       },
...
       "InvalidResponse":{
          "type":"object",
          "properties":{
             "statusCode":{
                "type":"string"
             },
             "message":{
                "type":"string"
             }
          }
       },
         "newCar": {
            "type": "object",
            "properties": {
              "selectCar": {
                "type": "object",
                "properties": {
                  "carBrand": {
                    "type": "string"
                  },
                  "version": {
                    "type": "string"
                  },
                  "makeYear": {
                    "type": "string"
                  }
                }
              },
              "carSpecs": {
                "type": "object",
                "properties": {
                  "AvalibleColors": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  },
                  "carfeaturs": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  },
                  "specs": {
                    "type": "string"
                  },
                  "CruiseControl": {
                    "type": "boolean"
                  },
                  "FuelType": {
                    "type": "string"
                  },
                  "EngineCapacity": {
                    "type": "string"
                  },
                  "BootCapacity": {
                    "type": "string"
                  }
                }
              },
              "PRICING": {
                "type": "object",
                "properties": {
                  "PricePerDay": {
                    "type": "string"
                  },
                  "MinimumDayBooking": {
                    "type": "string"
                  },
                  "PricePerWeek": {
                    "type": "string"
                  },
                  "ExtraMillingCost": {
                    "type": "string"
                  },
                  "CDW": {
                    "type": "string"
                  }
                }
              },
              "RentalTerms": {
                "type": "object",
                "properties": {
                  "Security": {
                    "type": "string"
                  },
                  "ExcessClaim": {
                    "type": "string"
                  },
                  "Delivery": {
                    "type": "string"
                  },
                  "SpecialNote": {
                    "type": "string"
                  }
                }
              },
              "MonthlyPricing": {
                "type": "object",
                "properties": {
                  "1month": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  },
                  "3months": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  },
                  "6months": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  },
                  "12months": {
                    "type": "array",
                    "items": {
                      "type": "integer"
                    }
                  },
                  "CDW": {
                    "type": "string"
                  }
                }
              }
            }
          },
...
"addContactInfo":{
   "type":"object",
   "properties":{
   "Email":{
      "type":"string"
   },
   "WhatsupNo":{
      "type":"string"
   },
   "SMSNotificationNo":{
      "type":"string"
   },
   "MainAdress":{
      "type":"string"
   },
   "MainAdressArabic":{
      "type":"string"
   },
   "PhoneNo":{
      "type":"string"
   }              
}

}
    }
 }